// src/types/job.types.ts
export interface JobData {
    title: string;
    description: string;
    companyName: string;
    location: string;
    minSalary: string;
    maxSalary: string;
    salary: string;
    eligibleYear: string;
    requiredExp: string;
    skills: string[];
    jobUrl: string;
    postPlatform: string;
    experienceLevel: string;
    position: string;
    postedAt: string | null;
    createdAt: string | null;
  }